requirements_docker.txt
aiocontextvars==0.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
contextvars==2.4
decorator==4.4.2
gTTS==2.2.3
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
immutables==0.16
importlib-metadata==4.8.1
loguru==0.5.3
moviepy==1.0.3
numpy==1.19.5
Pillow==8.3.1
proglog==0.1.9
pydub==0.25.1
pypiwin32==223
pyttsx3==2.90
pywin32==301
requests==2.26.0
six==1.16.0
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
win32-setctime==1.0.3
zipp==3.5.0
